fix tests

+++ src/maERP.Domain/Dtos/TaxClass/TaxClassDetailDto.cs
@@ -3,5 +3,5 @@
 public class TaxClassDetailDto
 {
     public int Id { get; set; }
-    public int TaxRate { get; set; }
+    public double TaxRate { get; set; }
 }
